Add a few tests, make some things private
